Convert project into a set of tabulation tools
Rather than coupling it to other programs as a C library, the project is
now converted into a set of tools for generating and processing files
containing matrix elements:

  - clh2of-tabulate: Calculate and store matrix elements in SQLite format
  - clh2of-pack: Pack the SQLite file into a native binary format
  - clh2of-unpack: Reverse of clh2of-pack
  - clh2of.py: A basic Python wrapper around the matrix element
    calculator; useful for testing and interactive evaluation

This way the matrix elements can be read in using standard tools (either
simple I/O, memory mapping, or SQLite), whichever is convenient.  The
SQLite format is arguably the easiest to use for a high-level language,
although it is less compact.